# Clean up apply! code