--- a/src/temporal.jl
+++ b/src/temporal.jl
@@ -5,5 +5,4 @@ Get the hour of day corresponding to the data.
 """
 struct HoD <: Transform end
 
-
 _apply(x, ::HoD; kwargs...) = hour.(x)