--- a/src/utils.jl
+++ b/src/utils.jl
@@ -11,14 +11,3 @@ function _try_copy(data)
         deepcopy(data)
     end
 end
-
-function invert_dims(A::AbstractArray, dims)
-    ndims(A) == 1 && return dims
-    # TODO: support named dims https://github.com/invenia/Transforms.jl/issues/20
-    inverted_dims = setdiff(1:ndims(A), dims)
-    if length(inverted_dims) == 1
-        inverted_dims = inverted_dims[1]
-    end
-
-    return inverted_dims
-end