Closes #18 

There is an unfortunate conflict in how the `dims` kwarg is implemented in `eachslice` vs `mapslices` #18 

`mapslices` is more intuitive to our researchers but it cannot be used to mutate data directly because it doesn't take a `view` of the underlying data (there is no such thing as a `mapview` https://github.com/JuliaLang/julia/issues/29146)

This is important for having a mutating `apply!` method, which is now only possible with `eachslices` + some workarounds.

This PR attempts to simplify this interface by just delegating to `apply` inside `apply!` and avoiding the complications involved in using `eachslice` in the first place. (The only place it is used in now is `LinearCombination`).

But since `apply!` delegates to `apply` which delegates to `_apply`, there is also no longer any need for the `_apply!` method.
So now users only need to implement one method (`_apply`) and `Transforms` takes care of the rest.

With this change I think the code is much simpler to follow and more "honest" in what it is using under-the-hood (`mapslices`).
However, I haven't taken steps to reduce allocations and it's probably 2x more expensive on that front but we can probably make those improvements as well.