@nicoleepp 2021-02-22
Probably worth going through the transforms and deleting the unnecessary `_apply!` methods