@glennmoy 2021-02-22
> Probably worth going through the transforms and deleting the unnecessary `_apply!` methods

yeah I think I got them all 👍