@glennmoy 2021-02-23
Some basic checks of performance
```julia
using Transforms
using Random
Random.seed!(1)

X = rand(1.0:10.0, 100, 100)

transforms = (
    Periodic(sin, 7),
    Power(3),
    MeanStdScaling(X; dims=1),
    LinearCombination(ones(100)),
    OneHotEncoding(1:10),
)

for t in transforms
    @time Transforms.apply(X, t; dims=1);
    @time Transforms.apply(X, t; dims=1);

    try
        @time Transforms.apply!(X, t; dims=1);
        X = rand(1.0:10.0, 100, 100)
        @time Transforms.apply!(X, t; dims=1);
    catch e
    end
end

"""
This branch:

Periodic(sin, 7)
apply
0.780758 seconds (2.84 M allocations: 146.639 MiB, 5.11% gc time)
0.000626 seconds (1.64 k allocations: 380.922 KiB)
apply!
0.027169 seconds (71.10 k allocations: 4.024 MiB)
0.000667 seconds (1.64 k allocations: 380.922 KiB)

Power(3)
apply
0.160279 seconds (347.52 k allocations: 18.862 MiB)
0.000445 seconds (1.74 k allocations: 221.500 KiB)
apply!
0.007062 seconds (5.40 k allocations: 417.305 KiB)
0.000387 seconds (1.74 k allocations: 221.500 KiB)

MeanStdScaling(X; dims=1)
apply
0.184251 seconds (419.53 k allocations: 22.622 MiB, 5.67% gc time)
0.000619 seconds (2.04 k allocations: 234.000 KiB)
apply!
0.007365 seconds (5.70 k allocations: 430.008 KiB)
0.000334 seconds (2.04 k allocations: 234.000 KiB)

LinearCombination(ones(100))
apply
0.329633 seconds (1.26 M allocations: 64.869 MiB, 3.28% gc time)
0.000814 seconds (30.61 k allocations: 812.000 KiB)

OneHotEncoding
apply
0.021401 seconds (31.42 k allocations: 1.939 MiB)
0.000913 seconds (11.98 k allocations: 968.531 KiB)
"""

"""
Main:

Periodic(sin, 7)
apply
0.675031 seconds (2.59 M allocations: 133.186 MiB, 4.47% gc time)
0.000668 seconds (1.64 k allocations: 463.734 KiB)
apply!
0.342571 seconds (1.51 M allocations: 78.363 MiB, 6.02% gc time)
0.000292 seconds (1.21 k allocations: 300.781 KiB)

Power(3)
apply
0.153669 seconds (338.18 k allocations: 18.510 MiB)
0.000476 seconds (1.54 k allocations: 291.812 KiB)
apply!
0.075057 seconds (210.79 k allocations: 10.918 MiB)
0.000199 seconds (1.61 k allocations: 155.469 KiB)

MeanStdScaling(X; dims=1)
apply
0.214341 seconds (575.51 k allocations: 31.068 MiB, 4.94% gc time)
0.000420 seconds (2.04 k allocations: 390.250 KiB)
apply!
0.093857 seconds (278.64 k allocations: 14.865 MiB)
0.000199 seconds (1.91 k allocations: 167.969 KiB)

LinearCombination(ones(100))
apply
0.308854 seconds (1.04 M allocations: 53.751 MiB, 6.56% gc time)
0.001061 seconds (30.61 k allocations: 812.000 KiB)

OneHotEncoding
apply
0.020939 seconds (31.30 k allocations: 1.934 MiB)
0.000875 seconds (11.98 k allocations: 968.531 KiB)
"""
```