@glennmoy 2021-02-23
> > Note that now apply and apply! are equally performant and are roughly the average of what they each were before (where apply got better and apply! got worse).
> 
> Hmm, true for the first timing, but for the second timing `apply` looks slightly worse on average. It's well within the same order of magnitude though, so it looks pretty good.

maybe I'm blind - can you point out where `apply` is doing worse in this branch?